requirements.txt
blinker
click
dnspython
Flask
Flask-PyMongo
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
Werkzeug
